cleanup utility doesn't remove file named "empty"
"empty" files are used to track empty folder in git and them must not be removed in clenup operations

+++ bedita-app/libs/be_system.php
@@ -171,10 +171,12 @@ class BeSystem {
 		// delete files
 		foreach ($list[1] as $file) {
 			$f = new File($file);
-			if (!$f->delete()) {
-				$results['failed'][] = array('error' => __("Error deleting file", true) . " " . $file);
-			} else {
-				$results['success'][] = $file . " " . __('deleted', true);
+			if ($f->name != "empty") {
+				if (!$f->delete()) {
+					$results['failed'][] = array('error' => __("Error deleting file", true) . " " . $file);
+				} else {
+					$results['success'][] = $file . " " . __('deleted', true);
+				}
 			}
 		}
 		// delete dirs
